+++ inst/scripts/pkgdown.sh
@@ -1,5 +1,8 @@
 #!/usr/bin/bash
 
+# convert -density 300 logo.svg logo.png
+# https://www.freeconvert.com/image-converter
+
 Rscript -e '
       knitr::knit("README.Rmd")
       library(pkgdown)