+++ NEWS.md
@@ -1,10 +1,23 @@
+## gaawr2 0.0.3
+
+(2025/3/24)
+
+* Add CRAN URL.
+* Amend README.Rmd/_pkgdown.yml after CRAN submission/publication.
+* Amend vignette which extracts a list of packages in `Suggests`.
+* Create inst/csl/ as flagged by CRAN & Prof Brian Ripley.
+* Set fig.height/fig.width in vignettes.
+* Suggests Cairo and use 'CairoPNG'.
+* Turn most inst/scripts/*R to be functions.
+
 ## gaawr2 0.0.2
 
-(2025/2/20)
+(2025/2/21)
 
 * Add index.md.
 * Expanding inst/scripts/ens.R so as to slim down pQTLdata.
 * Revise Title/DESCRIPTION and track options()'s as suggested by Beni Altmann.
+* Revise vignette.
 
 ## gaawr2 0.0.1